feat: add DeleteTaskControllerTest to validate task deletion functionality and authorization checks

+++ app/Http/Controllers/Api/V1/Tasks/DeleteTaskController.php
@@ -15,6 +15,11 @@ class DeleteTaskController extends Controller
      */
     public function __invoke(Task $task)
     {
+        if($task->created_by !== auth()->user()->id) {
+            return response()->json(['message' => 'You are not authorized to delete this task'], 
+                Response::HTTP_FORBIDDEN);
+        }
+        
         $task->delete();
         return response()->json(['message' => 'Task deleted successfully'], 
             Response::HTTP_OK);
